Fix the final-review findings in the pipeline
Money path (label stage): thinking is explicitly disabled on the models
whose default is adaptive thinking, max_tokens is 256, and a 50-request
canary batch must parse at 90% before the rest of the sample is
submitted (CanaryFailed, CLI exit 5). The batch id is written to the run
log the moment a batch is created so a Ctrl-C cannot orphan a paid
batch, and `loupe label --resume-batch ID` polls an existing one. A bad
result body costs one row instead of the whole batch, the labels parquet
is written in a finally, and a run that labels under 90% of the sample
is flagged not ok (CLI exit 4). Actual cost now counts cache tokens.

Metrics: return_rate is NULL for weeks whose following week is absent
from the data (with a next_week_present flag) instead of publishing 0.0,
which the dashboard would have read as churn; a new
pseudo_user_persistence_quarterly metric shows the pseudo-user key
losing week-to-week persistence around 2024-10. Country and language
volume files publish a suppressed_or_unknown residual row per week
(8.3% of the corpus for country) instead of dropping suppressed and
NULL cells. Slices group by model family, not by dated point release.
Intent coverage is "full" only when every flattened shard has
predictions. meta.json gains complete_weeks_from/through,
token_coverage_first_week, intent_share, the classifier report fields
and a caveats list. Sums are cast to BIGINT for Arrow JS, and coverage
rates no longer read -0.0.

+++ loupe/metrics/friction_weekly.sql
@@ -1,4 +1,4 @@
-SELECT week, count(*) AS conversations,
+SELECT week, count(*)::BIGINT AS conversations,
        avg(repeated_request::INT) AS repeat_rate, avg(one_and_done::INT) AS one_and_done_rate,
        avg(correction_followup::INT) AS correction_rate, avg(assistant_refusal::INT) AS refusal_rate
 FROM conversations WHERE NOT has_empty_user_input
